snd_unit: use ssize_t/size_t type for internal length representation

+++ src/internal.h
@@ -33,8 +33,7 @@ void hinawa_fw_resp_handle_request(HinawaFwResp *self,
 void hinawa_fw_req_handle_response(HinawaFwReq *self,
 				   struct fw_cdev_event_response *event);
 
-void hinawa_snd_unit_write(HinawaSndUnit *self,
-			   const void *buf, unsigned int length,
+void hinawa_snd_unit_write(HinawaSndUnit *self, const void *buf, size_t length,
 			   GError **exception);
 
 void hinawa_snd_unit_ioctl(HinawaSndUnit *self, unsigned long request,
@@ -42,23 +41,23 @@ void hinawa_snd_unit_ioctl(HinawaSndUnit *self, unsigned long request,
 
 #if HAVE_SND_DICE
 void hinawa_snd_dice_handle_notification(HinawaSndDice *self,
-					 const void *buf, unsigned int len);
+					 const void *buf, ssize_t len);
 #endif
 #if HAVE_SND_EFW
 void hinawa_snd_efw_handle_response(HinawaSndEfw *self,
-				    const void *buf, unsigned int len);
+				    const void *buf, ssize_t len);
 #endif
 #if HAVE_SND_DG00X
 void hinawa_snd_dg00x_handle_msg(HinawaSndDg00x *self, const void *buf,
-				 unsigned int len);
+				 ssize_t len);
 #endif
 #if HAVE_SND_MOTU
 void hinawa_snd_motu_handle_notification(HinawaSndMotu *self,
-					 const void *buf, unsigned int len);
+					 const void *buf, ssize_t len);
 #endif
 #if HAVE_SND_TSCM
 void hinawa_snd_tscm_handle_control(HinawaSndTscm *self, const void *buf,
-				    unsigned int len);
+				    ssize_t len);
 #endif
 
 #endif
